requirements.txt
amqp==5.0.9
aniso8601==7.0.0
asgiref==3.4.1
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2
django-environ==0.8.1
django-filter==21.1
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
graphene==2.1.9
graphene-django==2.15.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
kombu==5.2.3
packaging==21.3
Pillow==9.0.0
promise==2.3
prompt-toolkit==3.0.24
psycopg2==2.9.3
PyJWT==1.7.1
pyparsing==3.0.6
pytz==2021.3
redis==4.1.0
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
typing-extensions==4.0.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
